tablet: refactor delta iteration code
While examining existing delta iteration code, I noticed significant
similarities between the DeltaFileIterator and DMSIterator. After
shuffling some code around, some clear abstractions formed, and so this
patch introduces the DeltaStoreIterator that, along with the existing
DeltaPreparer, comprises implementations of the existing DeltaIterator.

The DeltaStoreIterator loads batches of deltas from the underlying
delta stores, to be passed to the DeltaPreparer and used as needed.
All existing implementations of DeltaIterators follow this pattern and
are updated to use the new abstraction as such.

I expect that in the future, this will also facilitate implementing
different styles of delta iteration, e.g. iterating over a delta store
as if all deltas were committed at a single timestamp, or applying the
deltas of several delta stores that overlap in time, etc.

Change-Id: Iafc9cffa51964f81bc01fd8fcf1b6a79770a7838